4.1.3 Добавит наследование простым компонентам
